nur so

+++ src/de/merit/azubi/Durchschnittsverbrauch.java
@@ -5,11 +5,12 @@ public class Durchschnittsverbrauch {
     public static void main(String[] args) {
 
         int gefKm = 430;
-        int verbrauchL = 38;
+        double verbrauchL = 38;
 
         double erg = verbrauchL / gefKm * 100;
 
         System.out.println (erg);
+
 
     }
 }
